fix(workspace): stabilize expert mention autocomplete
Extract the mention matching and interaction logic so the chat panel stays smaller while closing stale popovers on blur and keeping suggestions inside the viewport.
